restore test_appsync_events
